requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
black==21.12b0
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.4
click==8.0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
daphne==3.0.2
Django==3.2.6
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-cors-headers==3.8.0
django-js-asset==1.2.2
django-model-utils==4.1.0
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-gis==0.18
djangorestframework-simplejwt==4.8.0
drf-yasg==1.17.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mercadopago==2.0.8
msgpack==1.0.3
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
Pillow==8.2.0
platformdirs==2.4.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pytz==2021.1
qrcode==7.3.1
requests==2.26.0
ruamel.yaml==0.17.13
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
tomli==1.2.3
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.1
uritemplate==3.0.1
urllib3==1.26.6
zope.interface==5.4.0
